[refactor] 统一命名规范，优化时间戳类型，强化解码与线程逻辑
- 全模块替换调试打印为分模块日志输出
- PacketItem/FrameItem 时间戳改为 double 类型
- 统一函数命名：start/stop/run/init 等小写风格
- 重构 VideoConverter，支持动态分辨率切换
- 优化解码线程帧率控制，增加超时追帧修复
- 修复播放完毕后重启逻辑，避免线程死锁

+++ src/my_video_player/player/engine/common/engine_types.h
@@ -9,9 +9,9 @@ struct PacketItem {
 
     int stream_index = -1;
 
-    int64_t pts = AV_NOPTS_VALUE;
-    int64_t dts = AV_NOPTS_VALUE;
-    int64_t duration = 0;
+    double  pts = AV_NOPTS_VALUE;
+    double  dts = AV_NOPTS_VALUE;
+    double  duration = 0;
 
     int serial = 0;
 
@@ -30,8 +30,8 @@ struct FrameItem {
 
     int stream_index = -1;
 
-    int64_t pts = AV_NOPTS_VALUE;
-    int64_t duration = 0;
+    double  pts = AV_NOPTS_VALUE;
+    double  duration = 0;
 
     bool key_frame = false;
